+++ src/app/travel/reviews/detail/[id]/page.tsx
