+++ src/app/travel/reviews/page.tsx
@@ -4,6 +4,7 @@ import SearchBar from "@/components/SearchBar";
 import TripCard from "@/components/TripCard";
 import { recommandPlans } from "@/data/temporary";
 import Category from "@/components/Category";
+import CreatePostButton from "@/components/CreatePostButton";
 
 export const metadata: Metadata = {
   title: "Reviews",
@@ -17,7 +18,33 @@ export const metadata: Metadata = {
   },
 };
 
-export default function TravelReviewList() {
+async function getReviews() {
+  // try {
+  const headers = {
+    "Content-Type": "application/json",
+  };
+  const response = await fetch(
+    `${process.env.NEXT_PUBLIC_API_URL}/api/review`,
+    {
+      method: "GET",
+      headers,
+      cache: "no-store",
+    }
+  );
+  if (!response.ok) {
+    throw new Error("Failed to fetch reviews.");
+  }
+  const json = await response.json();
+  return { jsonData: json.data, accessToken: "null" };
+  // } catch (error) {
+  //   console.error("API 요청 중 오류 발생:", error);
+  //   return { jsonData: [], accessToken: "null" };
+  // }
+}
+
+export default async function TravelReviewList() {
+  const { jsonData, accessToken } = await getReviews();
+
   return (
     <main>
       <section className="page-section pt-4 pb-16">
@@ -26,8 +53,8 @@ export default function TravelReviewList() {
         <SearchBar />
         <Category page={"review"} />
         <section>
-          {/* <TripCard list={recommandPlans} page={"review"} /> */}
-          준비 중입니다.
+          <CreatePostButton page={"review"} />
+          <TripCard list={jsonData} accessToken={accessToken} page={"review"} />
         </section>
       </section>
     </main>
